refactor(network): extract multiaddr + IP rate-limit helpers from libp2p_node
The libp2p_node.rs file was past 1900 lines and the swarm logic
itself was getting hard to navigate around the small helpers. Two
sections were obviously self-contained:

- The multiaddr conversion stuff — `make_multiaddr` for legacy
  host:port → libp2p form, plus `extract_ip` for pulling v4/v6
  out of a ConnectedPoint. Pure functions, no swarm awareness.

- The per-IP connection rate limiter (`IpRateLimiter`). Standalone
  state, a check-and-track method, and a prune-stale janitor. Notable
  detail in the comment: the cap had to land at 20 connections per
  IP because the Treasury validator hosts a 5-validator stack on a
  single public address, and a fast-deploy rolling restart legitimately
  fires ~15 connection attempts from that one IP within a few seconds.

Pulled both into `libp2p_node/{multiaddr,ratelimit}.rs`. Public
API at `sentrix_network::libp2p_node::make_multiaddr` is preserved
via re-export; everything else stays `pub(super)`-scoped to the
parent module.

libp2p_node.rs drops 1986 → 1827 LOC. Test access to the limiter's
internal map went through a new `tracked_count()` helper rather
than poking the private field.

Live verified by cargo clippy + cargo test passing on the workspace.
